requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
python-dotenv
python-jose[cryptography]
python-multipart
authlib
httpx
itsdangerous